FE Userlist added
Finally loads list of users to FE and displays it
